# openapi.json
{"openapi":"3.0.1","info":{"title":"Beer Catalog API","description":"Beer search and save simple application","contact":{"name":"George I. Nicolae PFA","email":"george.nicolae.dev@gmail.com"},"version":"v0.0.1"},"servers":[{"url":"http://localhost:9090/beer-catalog"}],"tags":[{"name":"Beer Catalog API","description":"REST API that provides data about beer entries for a session"},{"name":"Beers Controller API","description":"Provides data about beer"}],"paths":{"/api":{"get":{"tags":["Beer Catalog API"],"summary":"Get beer catalog","operationId":"getBeerCatalog","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Beer catalog not found"},"200":{"description":"Beer catalog found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BeerCatalog"}}}}}},"post":{"tags":["Beer Catalog API"],"summary":"Create beer entry to catalog","operationId":"createBeerEntry","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BeerEntryDto"}}},"required":true},"responses":{"400":{"description":"Beer Entry Bad Request","content":{}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"string"}}}},"201":{"description":"Beer entry created","content":{"application/json":{}}}}}},"/beers":{"get":{"tags":["Beers Controller API"],"summary":"Get beers","operationId":"getBeers","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"Beers found","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Beer"}}}}},"404":{"description":"Beers not found"}}}},"/beers/{id}":{"get":{"tags":["Beers Controller API"],"summary":"Get beer by id","operationId":"getBeer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Beer not found"},"200":{"description":"Beer found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Beer"}}}}}}},"/beers/filter":{"get":{"tags":["Beers Controller API"],"summary":"Get beers filtered","operationId":"getBeersFiltered","parameters":[{"name":"params","in":"query","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"},"default":""}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"Beers found","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Beer"}}}}},"404":{"description":"Beers not found"}}}},"/api/{id}":{"get":{"tags":["Beer Catalog API"],"summary":"Get beer entry by id","operationId":"getBeerEntryById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Beer entry not found"},"200":{"description":"Beer entry found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BeerEntryDto"}}}}}}},"/api/search":{"get":{"tags":["Beer Catalog API"],"summary":"Get beer entry by name","operationId":"getBeerEntryByName","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Beer entry not found"},"200":{"description":"Beer entry found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BeerEntryDto"}}}}}}}},"components":{"schemas":{"BeerEntryDto":{"required":["beerId","beerName","location"],"type":"object","properties":{"beerId":{"type":"string"},"beerName":{"type":"string"},"location":{"type":"string"},"date":{"type":"string","format":"date"},"rating":{"maximum":5,"minimum":0,"type":"integer","format":"int32"},"comments":{"maxLength":255,"minLength":0,"type":"string"}}},"Beer":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"tagline":{"type":"string"},"firstBrewed":{"type":"string"},"description":{"type":"string"},"imageURL":{"type":"string"},"abv":{"type":"number","format":"double"},"ibu":{"type":"number","format":"double"},"targetFg":{"type":"number","format":"double"},"targetOg":{"type":"number","format":"double"},"ebc":{"type":"number","format":"double"},"srm":{"type":"number","format":"double"},"ph":{"type":"number","format":"double"},"attenuationLevel":{"type":"number","format":"double"},"volume":{"$ref":"#/components/schemas/BoilVolume"},"boilVolume":{"$ref":"#/components/schemas/BoilVolume"},"method":{"$ref":"#/components/schemas/Method"},"ingredients":{"$ref":"#/components/schemas/Ingredients"},"foodPairing":{"type":"array","items":{"type":"string"}},"brewersTips":{"type":"string"},"contributedBy":{"type":"string"}}},"BoilVolume":{"type":"object","properties":{"value":{"type":"number","format":"double"},"unit":{"type":"string"}}},"Fermentation":{"type":"object","properties":{"temp":{"$ref":"#/components/schemas/BoilVolume"}}},"Hop":{"type":"object","properties":{"name":{"type":"string"},"amount":{"$ref":"#/components/schemas/BoilVolume"},"add":{"type":"string"},"attribute":{"type":"string"}}},"Ingredients":{"type":"object","properties":{"malt":{"type":"array","items":{"$ref":"#/components/schemas/Malt"}},"hops":{"type":"array","items":{"$ref":"#/components/schemas/Hop"}},"yeast":{"type":"string"}}},"Malt":{"type":"object","properties":{"name":{"type":"string"},"amount":{"$ref":"#/components/schemas/BoilVolume"}}},"MashTemp":{"type":"object","properties":{"temp":{"$ref":"#/components/schemas/BoilVolume"},"duration":{"type":"integer","format":"int64"}}},"Method":{"type":"object","properties":{"mashTemp":{"type":"array","items":{"$ref":"#/components/schemas/MashTemp"}},"fermentation":{"$ref":"#/components/schemas/Fermentation"},"twist":{"type":"object"}}},"BeerCatalog":{"type":"object","properties":{"beerEntries":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/BeerEntryDto"}},"localDate":{"type":"string","format":"date"}}}}}}